confirm

+++ src/userprog/exception.c
@@ -159,7 +159,7 @@ page_fault (struct intr_frame *f)
   //printf ("user:%d\nnot_present:%d\nwrite:%d", user, not_present, write);
 
 
-  if (fault_addr != NULL && is_user_vaddr(fault_addr) && not_present && ((int)fault_addr > BOTTOM_UVADDR))
+  if (fault_addr != NULL && is_user_vaddr(fault_addr) && not_present && (fault_addr > BOTTOM_UVADDR))
   {
     bool can_load_page = false;
     bool can_grow_stack = false;
